+++ lib/segment/src/index/query_optimization/mod.rs
@@ -1,4 +1,3 @@
-pub mod condition_converter;
 pub mod optimized_filter;
 pub mod payload_provider;
 pub mod rescore_formula;
